[Grep]: Replace mock open with tempfiles.
Rather than mocking the `open()` built-in to provide the input files,
write the files to a TemporaryDirectory and change the current working
directory (CWD) to the tempdir.

This allows users to read the files using `open()`,
`pathlib.Path.read_text()` or any other IO method.

Writing the files incurs a small amount of disk IO. However, it
simplifies the test setup and removes the restriction on how the files
are read.

Steps taken:

* Modify the `template.j2` file.
* Run `./bin/generate_tests.py grep` to update the test file.
* Run `/bin/test_exercises.py grep` to verify the test (and example)
  work.

Additional steps:

* I changed my personal solution to use `pathlib.Path.read_text()` and
  verified the tests failed.
* I copied the generated test file to my personal Exercism workspace.
* I ran the newly generated test against my `pathlib` solution and
  verified the solution now works.
* I reverted my solution to use `open()` and verified it also still
  works against the newly generated tests.